refact!: Merge `Cms\Api` into `Api\Api` class
BREAKING CHANGE: Removed `Kirby\Cms\Api` class. Use `Kirby\Api\Api` class instead.
