+++ tests/Api/Collections/ChildrenCollectionTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Kirby\Cms;
+namespace Kirby\Api;
 
-use Kirby\Cms\Api\ApiCollectionTestCase;
+use Kirby\Cms\Site;
 
-class ChildrenApiCollectionTest extends ApiCollectionTestCase
+class ChildrenCollectionTest extends CollectionTestCase
 {
-	public const TMP = KIRBY_TMP_DIR . '/Cms.ChildrenApiCollection';
+	public const TMP = KIRBY_TMP_DIR . '/Api.ChildrenCollection';
 
 	public function testCollection(): void
 	{
