+++ tests/Api/Collections/LanguagesCollectionTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Kirby\Cms;
+namespace Kirby\Api;
 
-use Kirby\Cms\Api\ApiCollectionTestCase;
+use Kirby\Cms\App;
 use Kirby\Filesystem\Dir;
 
-class LanguagesApiCollectionTest extends ApiCollectionTestCase
+class LanguagesCollectionTest extends CollectionTestCase
 {
-	public const TMP = KIRBY_TMP_DIR . '/Cms.LanguagesApiCollection';
+	public const TMP = KIRBY_TMP_DIR . '/Api.LanguagesCollection';
 
 	public function setUp(): void
 	{
