+++ tests/Api/Collections/UsersCollectionTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Kirby\Cms;
+namespace Kirby\Api;
 
-use Kirby\Cms\Api\ApiCollectionTestCase;
+use Kirby\Cms\App;
 use Kirby\Filesystem\Dir;
 
-class UsersApiCollectionTest extends ApiCollectionTestCase
+class UsersCollectionTest extends CollectionTestCase
 {
-	public const TMP = KIRBY_TMP_DIR . '/Cms.UsersApiCollection';
+	public const TMP = KIRBY_TMP_DIR . '/Api.UsersCollection';
 
 	public function setUp(): void
 	{
