+++ tests/Api/Models/FileModelTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Kirby\Cms;
+namespace Kirby\Api;
 
-use Kirby\Cms\Api\ApiModelTestCase;
+use Kirby\Cms\Page;
 
-class FileApiModelTest extends ApiModelTestCase
+class FileModelTest extends ModelTestCase
 {
-	public const TMP = KIRBY_TMP_DIR . '/Cms.FileApiModel';
+	public const TMP = KIRBY_TMP_DIR . '/Api.FileModel';
 
 	public function testNextWithTemplate(): void
 	{
