+++ tests/Api/Models/PageModelTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Kirby\Cms;
+namespace Kirby\Api;
 
-use Kirby\Cms\Api\ApiModelTestCase;
+use Kirby\Cms\Page;
 
-class PageApiModelTest extends ApiModelTestCase
+class PageModelTest extends ModelTestCase
 {
-	public const TMP = KIRBY_TMP_DIR . '/Cms.PageApiModel';
+	public const TMP = KIRBY_TMP_DIR . '/Api.PageModel';
 
 	public function testChildren(): void
 	{
