+++ tests/Api/Routes/fixtures/avatar.jpg
